refactor(cli): update app.py and ops test imports (issue #4670 Part B)
- Update 6 import sites in cli/app.py (sessions_app, run_capture_store,
  run_codex_orphans, run_daemon_orphans, run_process_orphans,
  run_codex_attempts) to source from cli.ops
- Update 5 ops-cluster test files: test_capture_store,
  test_codex_attempts, test_codex_orphans_cmd, test_daemon_orphans_cli,
  test_process_orphans_cmd — both `import ... as` aliases and
  `from ... import` direct imports
