+++ src/autoskillit/cli/app.py
@@ -23,9 +23,9 @@ from autoskillit.cli._init_helpers import (
     _register_all,
 )
 from autoskillit.cli._serve_guard import serve_with_signal_guard
-from autoskillit.cli._sessions import sessions_app
 from autoskillit.cli._validate import validate_app
 from autoskillit.cli.fleet import fleet_app
+from autoskillit.cli.ops import sessions_app
 from autoskillit.cli.session._session_cook import cook as cook_interactive
 from autoskillit.cli.session._session_order import _recipes_dir_for, order
 from autoskillit.core import (
@@ -348,7 +348,7 @@ def doctor(*, output_json: bool = False):
 @app.command
 def capture_store(*, reclaim: bool = False) -> None:
     """Report capture-store ledger/directory statistics, or reclaim backlog with --reclaim."""
-    from autoskillit.cli._capture_store import run_capture_store
+    from autoskillit.cli.ops import run_capture_store
 
     run_capture_store(reclaim=reclaim)
 
@@ -356,7 +356,7 @@ def capture_store(*, reclaim: bool = False) -> None:
 @app.command
 def codex_orphans(*, reap: bool = False, output_json: bool = False) -> None:
     """Report orphaned codex TUI processes (deleted pty), or terminate them with --reap."""
-    from autoskillit.cli._codex_orphans import run_codex_orphans
+    from autoskillit.cli.ops import run_codex_orphans
 
     run_codex_orphans(reap=reap, output_json=output_json)
 
@@ -364,7 +364,7 @@ def codex_orphans(*, reap: bool = False, output_json: bool = False) -> None:
 @app.command
 def daemon_orphans(*, reap: bool = False, output_json: bool = False) -> None:
     """Report orphaned registered-stdio daemons, or terminate them with --reap."""
-    from autoskillit.cli._daemon_orphans import run_daemon_orphans
+    from autoskillit.cli.ops import run_daemon_orphans
 
     run_daemon_orphans(reap=reap, output_json=output_json)
 
@@ -372,7 +372,7 @@ def daemon_orphans(*, reap: bool = False, output_json: bool = False) -> None:
 @app.command
 def process_orphans(*, reap: bool = False, output_json: bool = False) -> None:
     """Report tethered children whose guardian is dead or ceiling passed, or reap with --reap."""
-    from autoskillit.cli._process_orphans import run_process_orphans
+    from autoskillit.cli.ops import run_process_orphans
 
     run_process_orphans(reap=reap, output_json=output_json)
 
@@ -385,7 +385,7 @@ def codex_attempts(
     output_json: bool = False,
 ) -> None:
     """List retained Codex attempt views, or explicitly discard one eligible view."""
-    from autoskillit.cli._codex_attempts import run_codex_attempts
+    from autoskillit.cli.ops import run_codex_attempts
 
     run_codex_attempts(
         discard_view=discard_view,
