+++ tests/cli/test_process_orphans_cmd.py
@@ -38,7 +38,7 @@ def _make_orphan(pid: int, reason: str = "spawner_dead") -> OrphanedTetherRecord
 
 
 def test_process_orphans_cmd_delegates(monkeypatch: pytest.MonkeyPatch) -> None:
-    import autoskillit.cli._process_orphans as process_orphans_mod
+    import autoskillit.cli.ops._process_orphans as process_orphans_mod
     from autoskillit import cli
 
     called_with: dict[str, object] = {}
@@ -58,7 +58,7 @@ def test_run_reports_orphans_plain(
     monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture[str]
 ) -> None:
     import autoskillit.execution as execution_mod
-    from autoskillit.cli._process_orphans import run_process_orphans
+    from autoskillit.cli.ops import run_process_orphans
 
     orphans = [_make_orphan(1001), _make_orphan(1002, reason="ceiling_expired")]
     monkeypatch.setattr(execution_mod, "find_orphaned_tethers", lambda _dir: orphans)
@@ -77,7 +77,7 @@ def test_run_reports_no_orphans(
     monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture[str]
 ) -> None:
     import autoskillit.execution as execution_mod
-    from autoskillit.cli._process_orphans import run_process_orphans
+    from autoskillit.cli.ops import run_process_orphans
 
     monkeypatch.setattr(execution_mod, "find_orphaned_tethers", lambda _dir: [])
 
@@ -90,7 +90,7 @@ def test_run_output_json_shape(
     monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture[str]
 ) -> None:
     import autoskillit.execution as execution_mod
-    from autoskillit.cli._process_orphans import run_process_orphans
+    from autoskillit.cli.ops import run_process_orphans
 
     orphans = [_make_orphan(1001), _make_orphan(1002)]
     monkeypatch.setattr(execution_mod, "find_orphaned_tethers", lambda _dir: orphans)
@@ -108,7 +108,7 @@ def test_run_reap_invokes_sweep_and_reports(
     monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture[str]
 ) -> None:
     import autoskillit.execution as execution_mod
-    from autoskillit.cli._process_orphans import run_process_orphans
+    from autoskillit.cli.ops import run_process_orphans
 
     orphans = [_make_orphan(2001), _make_orphan(2002)]
     monkeypatch.setattr(execution_mod, "find_orphaned_tethers", lambda _dir: orphans)
@@ -141,7 +141,7 @@ def test_run_reap_output_json_shape(
     monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture[str]
 ) -> None:
     import autoskillit.execution as execution_mod
-    from autoskillit.cli._process_orphans import run_process_orphans
+    from autoskillit.cli.ops import run_process_orphans
 
     orphans = [_make_orphan(3001)]
     monkeypatch.setattr(execution_mod, "find_orphaned_tethers", lambda _dir: orphans)
